+++ NEWS.md
@@ -21,6 +21,7 @@
   * Individuals can enter the model after the beginning with the `inflow` argument in `run_model()` (mainly for budget impact analysis).
   * Strategy name can be used to define values with `dispatch_strategy()` or using the `strategy` name (vignettes *homogeneous* and *probabilistic* have been updated to use this feature).
   * Beta and triangle distribution for PSA.
+  * Covariance analysis on strategy differences.
 
 ## Other features